tp-ci-cd/flask-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-flask==1.2.0